Consolidate auth and organization APIs
